mtd: api for mtd devices
Enable support of mtd devices and partitions on flash. Provide support
for mtd on eeprom and retained_ram.

Modified flash_map_default.c to avoid generating non flash partitions in
the default_flash_map structure.

Performed some optimisation in finding the mtd_dev_cfg and offset on the
device.

Reworked to make the read, write and erase functions on the device part
of the mtd_dev_cfg struct.

Signed-off-by: Laczen JMS <laczenjms@gmail.com>